requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
playwright==1.40.0
httpx==0.25.2
fake-useragent==1.4.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1
pydantic==2.5.2
jsonschema==4.20.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
optuna==3.5.0

# Feature Engineering
networkx==3.2.1
scipy==1.11.4
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
redis==5.0.1

# API
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2
sentry-sdk==1.39.1

# Utils
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
joblib==1.3.2
python-dateutil==2.8.2

# Data Validation
great-expectations==0.18.8
pandera==0.17.2

# Visualization (for notebooks)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Rate Limiting
ratelimit==2.2.1
tenacity==8.2.3
